requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.9
aiohttp_socks==0.9.0
aiosignal==1.3.1
amqp==5.2.0
Appium-Python-Client==4.1.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.8.0
blackbox==0.7.1
bleach==6.1.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-extensions==3.2.3
django-oauth2==4.2
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework~=3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
drf-yasg2==1.19.4
drf_api_logger==1.1.15
environ==1.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.65.5
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kombu==5.4.0
libclang==18.1.1
mailtrap
magiclog==1.0.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes
mtcnn
multidict
mypy-extensions
namex==0.0.8
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.0
pandas
pathspec==0.12.1
pep8
pillow
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psycopg
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.5.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
selenium==4.24.0
setuptools==72.2.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
stackclimber==0.99
termcolor==2.4.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
undetected-chromedriver==3.4.6
uritemplate==4.1.1
urllib3==2.2.1
v2==1.0.7
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.14.0

openai~=1.52.2
